8.0.1.0
Temporary fix so that "slider" frames will have a valid title, thus avoiding lua errors later on.

The edit at line 1482 takes advantage of automatically generated globals for every slider frame; but this is dependent on the slider having a unique name.

The edit at line 1899 is to ensure a unique name for every slider even if none was explicitly given by existing code.   A better solution would be to go through every other CT module and properly assign the name to those which lack it.... though such effort is not mutually exclusive of this edit so no need to refrain from committing this.